ADT/consumer/requirements.txt
kafka-python
tensorflow
pillow
